+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added support for Craft Commerce variants
+- Added support for Quick Post widgets
+
+### Fixed
+- Fixed a bug where batch actions bars weren't being initialised on Craft Commerce product edit pages
+
 ## 1.2.2 - 2023-02-23
 
 ### Fixed